Module krypto::authentication

🔑 Authentication

Username and password is used to authenticate users (humans).

Usernames are alphanumeric (letters A-Z, numbers 0-9) with the exception of underscores.

Password should be at least 10 characters long

The user’s password is hashed with scrypt and the hash is stored in the database.

The username may be claimed by a suspended or deactivated account. Suspended and deactivated usernames are not immediately available for use.

After the user has been authenticated, they are handed a passport that should send with requests to private resources.

kong allows a reserve list of usernames that can never be used by end-users (e.g admin)

Attaching to HTTP requests

Clients that request to access protected routes, need to provide a valid kpassport, they do this by attaching a kpassport with every request to a protected route. There are two ways to attach a kpassport to a request:

HTTP Cookie

The cookies Secure attribute is set, this ensures that the cookie is only sent over an HTTPS connection and not HTTP. This means that the cookie (kpassport) cannot be accessed by “MITM” attackers.

The cookies HttpOnly attribute is also set, this ensures that the cookie is inaccessible to the JavaScript Document.cookie API. So the cookie cannot be read or modified by client side JavaScript.

Cookie expiration date is also set. It is calculated from the kpassport’s timestamp:

TODO: how to calculate a cookies expiration date from a kpassport

Set-Cookie: session=; Expires=Thu, 21 Oct 2021 07:28:00 GMT; Secure; HttpOnly

Authorization header:

A kpassport can be transported using HTTP headers, the kpassport is sent in the Authorization header:

Authorization: Bearer <the kpassport kpassport>

Expiration

A kpassport is timestamped at the time it is issued

Structs

AuthHeaders

HTTP authentication methods

Enums

Auth

Kong authentication and authorization management